+++ CHANGES.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+# dbt-dry-run v0.6.7
+
+- Compatibility with dbt v1.5
+
+- Adds `--threads` option as an override
+
 # dbt-dry-run v0.6.6
 
 ## Bugfixes & Improvements